+++ app/Providers/AppServiceProvider.php
@@ -6,6 +6,7 @@ use Illuminate\Support\ServiceProvider;
 use Illuminate\Support\Facades\Gate;
 use App\Models\User;
 use App\Models\Project;
+use App\Models\Item;
 
 class AppServiceProvider extends ServiceProvider
 {
@@ -27,6 +28,9 @@ class AppServiceProvider extends ServiceProvider
         });
         Gate::define('edit-project', function(User $user, Project $project){
             return auth()->user()->is_admin === 1 || $project->user_id == auth()->id();
+        });
+        Gate::define('edit-item', function(User $user, Item $item){
+            return auth()->user()->is_admin === 1 || $item->user_id == auth()->id();
         });
     }
 }