+++ resources/views/items/form.blade.php
@@ -18,15 +18,17 @@
             @endif
             <input type="hidden" name="project_id" value="{{ $item->project->id }}" />
 
-            <x-text-input name="title" class="w-full" placeholder="Titel" />
-            <x-input-error :messages="$errors->title->get('title')" />
+            <x-text-input name="title" class="w-full" placeholder="Titel" value="{{ old('title', $item->title) }}"/>
+            <x-input-error :messages="$errors->get('title')" />
 
             <x-forms.tinymce-editor name="story">{{ old('story', $item->story) }}</x-forms.tinymce-editor>
 
             <x-h2 class="my-4">Typ</x-h2>
+            <x-input-error :messages="$errors->get('type')" />
             <x-item-type name="type" :value="$item->type"/>
 
             <x-h2 class="my-4">Priorität</x-h2>
+            <x-input-error :messages="$errors->get('priority')" />
             <x-item-priority name="priority" :value="$item->priority"/>
 
             <x-primary-button class="mt-4">Speichern</x-primary-button>