+++ resources/views/projects/show.blade.php
@@ -39,14 +39,14 @@
                     <ul class="flex gap-4 items-center">
                         <li>{!! $item->typePillHtml() !!}</li>
                         <li>{!! $item->priorityPillHtml() !!}</li>
-                        <li class="text-gray-500 text-sm">{{ $item->created_at?->format('d.m.Y H:i') }}</li>
+                        <li class="text-gray-500 text-sm">{{ formatDateTime($item->created_at) }}</li>
                     </ul>
                 </div>
                 <div class="pt-2">
                     <p>{{ $item->comments()->count() }} Kommentare</p>
                     <a href="{{ route('items.show', ['project' => $project->id, 'item' => $item->id]) }}" class="text-sm flex gap-2 items-center font-medium text-gray-600">Details ansehen <x-icons.right-arrow width="1em" height="1em"/></a>
 
-                    @can('admin')
+                    @can('edit-item', $item)
                         <div class="mt-4">
                             <form action="{{ route('items.destroy', ['project' => $project->id, 'item' => $item->id]) }}" method="POST">
                                 @csrf