+++ CHANGELOG.md
@@ -4,6 +4,12 @@ Updates should follow the [Keep a CHANGELOG](https://keepachangelog.com/) princi
 
 ## [Unreleased]
 
+## [1.0.1] - 29.01.2025
+
+### Added
+- Upvote error and success messages
+- Allow deleting own items
+
 ## [1.0.0] - 29.01.2025
 
 ### Added